moving trades up to parent
