+++ src/components/Form.jsx
@@ -6,11 +6,12 @@ import { collection, doc, query, addDoc, getDocs, serverTimestamp, where, limit,
 
 const Form = (props) => {    
 
-    const { auth, db } = props
+    const { auth, db, trades } = props
 
     const [value, setValue] = useState(0)
     const [bidOrAsk, setBidOrAsk] = useState('')
     // const [resting, setResting] = useState(0)
+    const bids = trades.map(trades => trades[0]);
     // const [counterParty, setCounterParty] = useState('')
 
     async function postTrade() {
@@ -46,23 +47,39 @@ const Form = (props) => {
       
       
       const isBid = bidOrAsk == 'bid';
-      
-      
-      const q = query(collection(db, "orders"), where("bidOrAsk", "==", ((isBid) ? "ask" : "bid")), orderBy('value', ((isBid) ? "asc" : "desc")), limit(1));
-      const querySnapshot = await getDocs(q);
-      if (querySnapshot.docs.length != 0) {
 
-        querySnapshot.forEach((doc) => {
-          ((isBid && doc.data().value && value >= doc.data().value) || !isBid && doc.data().value && value <= doc.data().value) ?
-          (updateParties(doc.data().uid, doc.data().value, isBid, doc.ref)) :
+      const onesideds = bids.filter(item => item.bidOrAsk === (isBid ? 'ask' : 'bid'));
+      const index = (isBid ? 0 : onesideds.length-1)
+      const bestOffer = onesideds[index].value
+      const docRef = trades[bids.indexOf(onesideds[index])][1]
+      console.log(docRef);
+      
+      // console.log(bids)
+      (onesideds.length != 0) ?
+      (
+        // console.log(bestOffer)
+        ((isBid && bestOffer && value >= bestOffer) || !isBid && bestOffer && value <= bestOffer) ?
+          (updateParties(onesideds[index].uid, bestOffer, isBid, docRef)) :
           (postTrade())
-
-          console.log(doc.data().value)
-        });
-        
-      } else {
+      ) : 
+      (
         postTrade()
-      }
+      )
+      console.log(onesideds[index].value);
+      // const q = query(collection(db, "orders"), where("bidOrAsk", "==", ((isBid) ? "ask" : "bid")), orderBy('value', ((isBid) ? "asc" : "desc")), limit(1));
+      // const querySnapshot = await getDocs(q);
+      // if (querySnapshot.docs.length != 0) {
+
+      //   querySnapshot.forEach((doc) => {
+      //     ((isBid && doc.data().value && value >= doc.data().value) || !isBid && doc.data().value && value <= doc.data().value) ?
+      //     (updateParties(doc.data().uid, doc.data().value, isBid, doc.ref)) :
+      //     (postTrade())
+
+      //   });
+
+      // } else {
+      //   postTrade()
+      // }
       
       setValue(0)
       setBidOrAsk('')
